requirements.txt
Django==5.1.2
djangorestframework==3.14.0